Added check if the file has been processed
This check is added to prevent the user from opening multiple websocket connections to the url they are redirected to to download the result. The user can either download the result or if he visits this url again, he will be redirected to the main page. Because files are removed from storage after processing.